usm: tests: Move tests to OS specific files
